+++ test/unit/mdb/test_mdb_shell.cpp
@@ -48,15 +48,13 @@ TEST(mdb_shell, empty_lines)
   ASSERT_TRUE(d.call_graphs().empty());
 
   sh.line_available(data::user_input(" "), d, h);
-  ASSERT_EQ(std::vector<data::user_input>{data::user_input(" ")}, h.commands());
+  ASSERT_EQ(std::vector<data::user_input>{}, h.commands());
   ASSERT_EQ(data::empty_container(), d.raw_texts());
   ASSERT_EQ(data::empty_container(), d.types());
   ASSERT_TRUE(d.call_graphs().empty());
 
   sh.line_available(data::user_input("\t"), d, h);
-  ASSERT_EQ((std::vector<data::user_input>{
-                data::user_input(" "), data::user_input("\t")}),
-            h.commands());
+  ASSERT_EQ((std::vector<data::user_input>{}), h.commands());
   ASSERT_EQ(data::empty_container(), d.raw_texts());
   ASSERT_EQ(data::empty_container(), d.types());
   ASSERT_TRUE(d.call_graphs().empty());
@@ -99,21 +97,16 @@ TEST(mdb_shell, identical_all_space_lines_in_history)
   ASSERT_EQ(data::empty_container(), h.commands());
 
   sh.line_available(data::user_input(" "), d, h);
-  ASSERT_EQ(std::vector<data::user_input>{data::user_input(" ")}, h.commands());
+  ASSERT_EQ(std::vector<data::user_input>{}, h.commands());
 
   sh.line_available(data::user_input(" "), d, h);
-  ASSERT_EQ(std::vector<data::user_input>{data::user_input(" ")}, h.commands());
+  ASSERT_EQ(std::vector<data::user_input>{}, h.commands());
 
   sh.line_available(data::user_input("  "), d, h);
-  ASSERT_EQ((std::vector<data::user_input>{
-                data::user_input(" "), data::user_input("  ")}),
-            h.commands());
+  ASSERT_EQ((std::vector<data::user_input>{}), h.commands());
 
   sh.line_available(data::user_input(" "), d, h);
-  ASSERT_EQ((std::vector<data::user_input>{data::user_input(" "),
-                                           data::user_input("  "),
-                                           data::user_input(" ")}),
-            h.commands());
+  ASSERT_EQ((std::vector<data::user_input>{}), h.commands());
 }
 
 TEST(mdb_shell, skips_empty_lines)