+++ test/unit/mdb/test_readme.cpp
@@ -26,9 +26,11 @@ using namespace metashell;
 
 TEST(readme, continue_abbreviated_as_c)
 {
-  const data::mdb_command c("c");
+  const boost::optional<data::mdb_command> c = data::mdb_command::parse("c");
+  ASSERT_TRUE(c);
+
   const mdb::command command =
-      get_command_from_map(mdb::shell::build_command_handler(false), c);
+      get_command_from_map(mdb::shell::build_command_handler(false), *c);
   const auto keys = command.get_keys();
 
   ASSERT_TRUE(std::find(keys.begin(), keys.end(),