+++ docs/about/changelog.md
@@ -14,6 +14,7 @@
     [ECMAScript](https://en.cppreference.com/w/cpp/regex/ecmascript).
     * The default engine is `auto`.
     * Improved error messages for invalid mdb commands.
+    * Improved input validation
 
 ## Version 4.0.0